requirements.txt
alembic==1.9.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
requests==2.28.2
SQLAlchemy==1.4.46
urllib3==1.26.14
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
